Fixes to ruby client.
Fix issue with base client module and import name
Add a single hash argument for the request body for requests that have body parameters (required or not). Also tried to cleanup the method documentation.
Add basePath in the yaml specification (contextPath in code) to the base URL in the client.